pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "creating a scalable rag system"
authors = [
    {name = "Amal Thomas",email = "reachthomas98@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"

packages = [{ include = "rag", from = "src"}]

dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "openai (>=2.24.0,<3.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "pypdf (>=6.7.5,<7.0.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "multipart (>=1.3.1,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
